+++ src/compute-engine/numeric-value/big-numeric-value.ts
@@ -226,10 +226,28 @@ export class BigNumericValue extends NumericValue {
   }
 
   mul(other: number | Decimal | NumericValue): NumericValue {
-    if (this.isZero) return this;
+    if (this.isZero) {
+      if (
+        other instanceof NumericValue &&
+        (other.isPositiveInfinity ||
+          other.isNegativeInfinity ||
+          other.isComplexInfinity ||
+          other.isNaN)
+      )
+        return this._makeExact(NaN);
+      return this;
+    }
     if (other === 1) return this;
     if (other === -1) return this.neg();
-    if (other === 0) return this.clone(0);
+    if (other === 0) {
+      if (
+        this.isPositiveInfinity ||
+        this.isNegativeInfinity ||
+        this.isComplexInfinity
+      )
+        return this._makeExact(NaN);
+      return this.clone(0);
+    }
 
     if (this.isOne) {
       if (typeof other === 'number' || other instanceof Decimal)
@@ -259,7 +277,15 @@ export class BigNumericValue extends NumericValue {
     }
     if (other.isOne) return this;
     if (other.isNegativeOne) return this.neg();
-    if (other.isZero) return this.clone(0);
+    if (other.isZero) {
+      if (
+        this.isPositiveInfinity ||
+        this.isNegativeInfinity ||
+        this.isComplexInfinity
+      )
+        return this._makeExact(NaN);
+      return this.clone(0);
+    }
 
     if (this.im === 0 && other.im === 0)
       return this.clone(this.decimal.mul(other.bignumRe ?? other.re));