+++ test/compute-engine/latex-syntax/arithmetic.test.ts
@@ -65,8 +65,8 @@ describe('SUM parsing', () => {
       [
         "Sum",
         ["Add", "j", "k"],
-        ["Limits", "k", ["Add", "a", ["Power", 2, 3]], ["Square", 3]],
-        ["Limits", "j", ["Add", 2, ["Square", 3]], "b"]
+        ["Limits", "k", ["Add", "a", 8], 9],
+        ["Limits", "j", 11, "b"]
       ]
     `);
   });